Fixed the jumping problem, as well as button
You could jump while still not entering collision which ended up stopping jumping from happening again until you ran into something else. The door was connected to two buttons, both simultaniously using updates with arguments, if one of the arguments wasn't supplied to, nothing would happen, it was an and statement but with a door in the middle.